requirements.txt
click==8.1.7
fastapi==0.109.0
hydra-core==1.3.2
pandas==2.1.4
protobuf==4.25.2
pytest==7.4.4
scikit_learn==1.3.2
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
wandb==0.16.2
google-cloud-storage
